+++ HTauTauTreeBase.h
@@ -48,8 +48,8 @@ public :
   TLorentzVector getGenComponentP4(unsigned int index, unsigned int iAbsCharge);
 
   template<typename T> T getBranchValue(char *branchAddress, unsigned int index);
-  float getProperty(std::string name, unsigned int index);
-  std::vector<float> getProperties(const std::vector<std::string> & propertiesList, unsigned int index);
+  long double getProperty(std::string name, unsigned int index);
+  std::vector<long double> getProperties(const std::vector<std::string> & propertiesList, unsigned int index);
   void writePropertiesHeader(const std::vector<std::string> & propertiesList);
   void writeTriggersHeader(const TH1F*);
